Access level additions.

+++ CondemnedAssistance/Controllers/RegisterController.cs
@@ -1,5 +1,5 @@
-﻿using CondemnedAssistance.Models;
-using CondemnedAssistance.Services.Requirements;
+﻿using CondemnedAssistance.Helpers;
+using CondemnedAssistance.Models;
 using CondemnedAssistance.ViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -12,23 +12,37 @@ namespace CondemnedAssistance.Controllers {
     public class RegisterController : Controller {
         private UserContext _db;
         private IAuthorizationService _authorizationService;
+        private RegisterHelper _registerHelper;
 
         public RegisterController(UserContext context, IAuthorizationService authorizationService) {
             _db = context;
             _authorizationService = authorizationService;
+            _registerHelper = new RegisterHelper(context);
         }
 
         [HttpGet]
         public IActionResult Index() {
             List<RegisterModel> registers = new List<RegisterModel>();
-            _db.Registers.ToList().ForEach(r => {
-                RegisterLevelModel registerLevelModel = new RegisterLevelModel();
-                RegisterLevel registerLevel = _db.RegisterLevels.FirstOrDefault(row => row.Id == r.RegisterLevelId);
-                registerLevelModel.Id = registerLevel.Id;
-                registerLevelModel.Name = registerLevel.Name;
-                registerLevelModel.Description = registerLevel.Description;
-                registers.Add(new RegisterModel { Id = r.Id, Name = r.Name, Description = r.Description, RegisterLevelId = r.RegisterLevelId, RegisterLevels = new List<RegisterLevelModel> { registerLevelModel } });
-            });
+            if (User.IsInRole("3")) {
+                _db.Registers.ToList().ForEach(r => {
+                    RegisterLevelModel registerLevelModel = new RegisterLevelModel();
+                    RegisterLevel registerLevel = _db.RegisterLevels.FirstOrDefault(row => row.Id == r.RegisterLevelId);
+                    registerLevelModel.Id = registerLevel.Id;
+                    registerLevelModel.Name = registerLevel.Name;
+                    registerLevelModel.Description = registerLevel.Description;
+                    registers.Add(new RegisterModel {
+                        Id = r.Id,
+                        Name = r.Name,
+                        Description = r.Description,
+                        RegisterLevelId = r.RegisterLevelId,
+                        RegisterLevels = new List<RegisterLevelModel> { registerLevelModel } });
+                });
+            } else {
+                int registerId = Convert.ToInt32(HttpContext.User.FindFirst(c => c.Type == "RegisterId").Value);
+                int userId = Convert.ToInt32(HttpContext.User.Identity.Name);
+                registers = _registerHelper.GetUserRegisterModels(userId, registerId);
+            }            
+            
             return View(registers);
         }
 
@@ -41,11 +55,16 @@ namespace CondemnedAssistance.Controllers {
                 return new ChallengeResult();
             }
 
+            int[] registerChildren = _registerHelper.GetRegisterChildren(new int[] { }, Convert.ToInt32(HttpContext.User.FindFirst(c => c.Type == "RegisterId").Value));
             Register parentRegister = _db.Registers.FirstOrDefault(r => r.Id == parentId);
             RegisterModel model = new RegisterModel();
             List<RegisterLevelModel> registerLevels = new List<RegisterLevelModel>();
             _db.RegisterLevels.ToList().ForEach(row => {
-                registerLevels.Add(new RegisterLevelModel { Id = row.Id, Name = row.Name, Description = row.Description });
+                registerLevels.Add(new RegisterLevelModel {
+                    Id = row.Id,
+                    Name = row.Name,
+                    Description = row.Description
+                });
             });
             model.RegisterLevels.AddRange(registerLevels);
             model.RegisterLevelId = levelId;
@@ -159,21 +178,30 @@ namespace CondemnedAssistance.Controllers {
 
             Dictionary<string, int> registerActions = new Dictionary<string, int>();
             registerActions.Add("levelId", register.RegisterLevelId);
-            //int[] children = getRegisterChildren(id, new int[] { });
+            int[] children = _registerHelper.GetRegisterChildren(new int[] { }, id);
             if (!await _authorizationService.AuthorizeAsync(User, registerActions, "resource-register-actions-policy")) {
                 return new ChallengeResult();
             }
 
-            //if (children != null) {
-            //    ModelState.AddModelError("", "Register has children so it cannot be deleted");
-            //    return RedirectToAction("Index", "Register");
-            //}
-            
+            if (_db.UserRegisters.Any(u => u.RegisterId == id)) {
+                ModelState.AddModelError("", "Register has users so it cannot be deleted");
+                return RedirectToAction("Index", "Register");
+            }
+
+            if (children != null) {
+                ModelState.AddModelError("", "Register has children so it cannot be deleted");
+                return RedirectToAction("Index", "Register");
+            }
+
+            var parents =  _db.RegisterHierarchies.Where(r => r.ChildRegister == id).ToArray();
+            _db.RegisterHierarchies.RemoveRange(parents);
+
             _db.Registers.Remove(register);
             _db.SaveChanges();
             return RedirectToAction("Index", "Register");
         }
 
+        [Authorize(Roles = "3")]
         [HttpGet]
         public IActionResult RegisterLevels() {
             List<RegisterLevelModel> registerLevels = new List<RegisterLevelModel>();
@@ -187,11 +215,13 @@ namespace CondemnedAssistance.Controllers {
             return View(registerLevels);
         }
 
+        [Authorize(Roles = "3")]
         [HttpGet]
         public IActionResult CreateRegisterLevel() {
             return View();
         }
 
+        [Authorize(Roles = "3")]
         [HttpPost]
         public IActionResult CreateRegisterLevel(RegisterLevelModel model) {
             if (ModelState.IsValid) {
@@ -215,6 +245,7 @@ namespace CondemnedAssistance.Controllers {
             return  View(model);
         }
 
+        [Authorize(Roles = "3")]
         [HttpGet]
         public IActionResult UpdateRegisterLevel(int id) {
             RegisterLevel registerLevel = _db.RegisterLevels.FirstOrDefault(r => r.Id == id);
@@ -232,6 +263,7 @@ namespace CondemnedAssistance.Controllers {
             return View(model);
         }
 
+        [Authorize(Roles = "3")]
         [HttpPost]
         public IActionResult UpdateRegisterLevel(int id, RegisterLevelModel model) {
             if (ModelState.IsValid) {
@@ -251,6 +283,7 @@ namespace CondemnedAssistance.Controllers {
             return View(model);
         }
 
+        [Authorize(Roles = "3")]
         [HttpGet]
         public IActionResult DeleteRegisterLevel(int id) {
             if (_db.Registers.Count(r => r.RegisterLevelId == id) > 0) {
